requirements.txt
astroid==2.2.2
bleach==2.1.2
certifi==2017.11.5
chardet==3.0.4
Django==2.2.5
django-braces==1.8.1
django-contrib-comments==1.8.0
django-cors-headers==3.1.0
django-filter==2.4.0
django-rest-swagger==0.3.5
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
docutils==0.14
drfdocs==0.0.9
future==0.16.0
idna==2.6
importlib-metadata==2.1.0
isort==4.3.4
jmespath==0.9.3
JPype1==0.6.2
lazy-object-proxy==1.3.1
Markdown==3.2.2
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.6
pylint-plugin-utils==0.5
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.11
requests==2.18.4
requests-oauthlib==0.8.0
shortuuid==0.4.3
six==1.11.0
sqlparse==0.3.0
typed-ast==1.2.0
tzlocal==1.5.1
urllib3==1.22
webencodings==0.5.1
wrapt==1.11.1
zipp==1.2.0
